packages/services/pyproject.toml
[project]
name = "services"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.19.0",
    "azure-communication-email>=1.0.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"
